debt: playground: extract device-config + history-graph from main.js
Two more behavior-preserving moves. Both use ESM live bindings for
the mutable vars (timeSeriesStore, graphRange, showAllSensors) that
main.js still owns — the history-graph module imports them
directly and sees reassignments in place, so no getter plumbing
was needed.

- main/device-config.js: init + load/save + form population of the
  /api/device-config form. Calls into watchdog-ui for mode-
  enablement rendering.
- main/history-graph.js: drawHistoryGraph, collectSeriesPts,
  drawTempLine, tankAvgOf, toSchematicState. `export` dropped from
  the helpers knip flagged as internal-only.

main.js: 2716 → 2328 lines. npm test:unit 782/782 pass. knip clean.
